Add regression test
