+++ mock-sequencer/index.ts
@@ -7,19 +7,19 @@ const sequencer = new MockSequencer()
 const app = express()
 app.use(express.json())
 
-app.get('/skip-next', (_req, res) => {
+app.post('/skip-next', (_req, res) => {
   const count = sequencer.getCurrentCount()
   sequencer.setSkipNext()
   res.json(count)
 })
 
-app.get('/send-in-random', (_req, res) => {
+app.post('/send-in-random', (_req, res) => {
   const count = sequencer.getCurrentCount()
   sequencer.setSendInRandom()
   res.json(count)
 })
 
-app.get('/send-oversized', (_req, res) => {
+app.post('/send-oversized', (_req, res) => {
   const count = sequencer.getCurrentCount()
   sequencer.setSendOversized()
   res.json(count)
@@ -30,7 +30,7 @@ app.get('/block-number', (_req, res) => {
   res.json(count)
 })
 
-app.get('/reset', (_req, res) => {
+app.post('/reset', (_req, res) => {
   sequencer.reset()
   res.json(sequencer.getCurrentCount())
 })
